Version 1.5.1
When computing the number of bits of a int, the code is now CHAR_BIT*sizeof(int) rather than 8*sizeof(int), since its possible that the a char could have more than 8 bits on some machines.

Added code to the bitfieldspecial, scaledencode/scaleddecode, and fieldencode/fielddecode outputs to protect any 64 bit code with #ifdef UINT64_MAX. This should make it easier to use a common set of helper files across multiple architectures which do and do not support 64 bit types.