Version 1.5.3.
Created an XML parser to allow ProtoGen to locate elements in the .xml file for error reporting (sucks that Qt doesn't have this feature, I feel stupid writing it myself). Lots of reorganization on path management. ProtoGen no longer changes the process working directory, instead paths are managed internally. This way error reporting can report the correct relative path to files. Updated error reporting to include the line number of the xml file where the error was generated. You can now include ProtoGen as a prebuild step and the IDE (tested in Qt Creator) will correctly interpret the warnings and errors, allowing you to click on the warning to go directly to the line in the xml that is causing problems.